Changes following community feedback
Changes have been made to the proposal based on the feedback from the community of governors. In this revised proposal, the project scope has been reduced, and the roadmap and fund allocation have been clarified. Additionally, the cost of the first milestone has been decreased with the support of AF.