Create xgov-112.md
We're back! In this proposal we want to add in NFD Vault Support to reduce the opt-friction for new Algorand Users, Add ARC-36 Support (Non-Rarity Filters) to our Simple Mint and Update Tools, Add in advanced asset tools to change the Manager, Freeze, and Clawback Address, and add in a tool where you can enter a Wallet Address to download the tx history. In addition we plan on updating the site with a redesign.